pyproject.toml
[project]
name = "projmgt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "chromadb>=1.5.8",
  "cryptography>=46.0.7",
  "django>=6.0.4",
  "django-click>=2.5.0",
  "django-ninja>=1.6.2",
  "django-stubs>=6.0.3",
  "dspy>=3.1.3",
  "httpx>=0.28.1",
  "mem0ai[nlp]>=2.0.0",
  "mlflow>=2.18.0",
  "model-bakery>=1.23.4",
  "pydantic-settings>=2.14.0",
  "pytest>=9.0.3",
  "pytest-django>=4.12.0",
  "python-dotenv>=1.0.1",
  "rich>=15.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = ["."]
testpaths = ["tests"]
markers = ["live_llm: tests that call a real language model"]

[dependency-groups]
dev = ["pyright>=1.1.409"]

[tool.ruff]
line-length = 88
